+++ constructorio-client/src/test/java/io/constructor/client/ConstructorTest.java
@@ -1,7 +1,6 @@
 package io.constructor.client;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 import java.util.*;
@@ -48,43 +47,6 @@ public class ConstructorTest {
         assertEquals("autocomplete query should return something", "[]", res.toString());
     }
 
-    @Test
-    public void addNoParamsShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        assertTrue("addition without params succeeds", constructor.add(randStr, "Search Suggestions"));
-    }
-
-    @Test
-    public void addWithParamsShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        HashMap<String, Object> params = new HashMap<String, Object>();
-        params.put("suggested_score", 1337);
-        assertTrue("addition with params succeeds", constructor.add(randStr, "Search Suggestions", params));
-    }
-
-    @Test
-    public void addOrUpdateShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        assertTrue("upsert succeeds", constructor.addOrUpdate(randStr, "Search Suggestions"));
-    }
-
-    @Test
-    public void addConstructorItemShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        assertTrue("addition of a constructor item succeeds", constructor.add(new ConstructorItem(randStr), "Search Suggestions"));
-    }
-
-    @Test
-    public void creatingConstructorItemWithNullNameShouldFail() throws Exception {
-        thrown.expect(IllegalArgumentException.class);
-        HashMap<String, Object> params = new HashMap<String, Object>();
-        new ConstructorItem(null, params);
-    }
-
     /***
      * Tests addBatch(ConstructorItem[] items, String autocompleteSection)
      *
@@ -114,79 +76,10 @@ public class ConstructorTest {
         assertTrue("batch upsert succeeds", constructor.addOrUpdateBatch(items, "Search Suggestions"));
     }
 
-    @Test
-    public void addConstructorItemWithHashMapParamsShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        HashMap<String, Object> params = new HashMap<String, Object>();
-        params.put("suggested_score", 1337);
-        assertTrue("addition of a constructor item with params as a HashMap succeeds", constructor.add(new ConstructorItem(randStr, params), "Search Suggestions"));
-    }
-
-    @Test
-    public void addConstructorItemWithFunctionParamsShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        assertTrue("addition of a constructor item with params set using the functions provided succeeds", constructor.add(new ConstructorItem(randStr).setSuggestedScore(100), "Search Suggestions"));
-    }
-
-    @Test
-    public void addConstructorItemWithEdgeCasesShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        HashMap<String, Object> params = new HashMap<String, Object>();
-        params.put("id", "test");
-        assertTrue("addition of a constructor item with some edge cases succeeds", constructor.add(new ConstructorItem("")
-                        .setSuggestedScore(1009)
-                        .setSuggestedScore(0)
-                        .setSuggestedScore(100)
-                        .setKeywords("hello", "world")
-                        .put("keywords", null)
-                        .put("item_name", randStr)
-                , "Search Suggestions"));
-    }
-
-    @Test
-    public void constructorItem() throws Exception {
-        String randStr = this.getRandString();
-        HashMap<String, Object> params = new HashMap<String, Object>();
-        params.put("id", "test");
-        assertEquals("ConstructorItem handles a missing suggested_score correctly", new ConstructorItem(randStr, params)
-                .setSuggestedScore(1009)
-                .setSuggestedScore(0)
-                .get("suggested_score"), null);
-        assertEquals("ConstructorItem handles constructor params correctly", new ConstructorItem(randStr, params)
-                .get("id"), "test");
-        assertEquals("ConstructorItem handles keywords correctly", new ConstructorItem(randStr, params)
-                .setKeywords("hello", "world")
-                .getKeywords()[1], "world");
-        assertNull("ConstructorItem handles item removal by setting the value to null correctly", new ConstructorItem(randStr, params)
-                .setKeywords(null)
-                .getKeywords());
-    }
-
-    @Test
-    public void removeShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        constructor.add(randStr, "Search Suggestions");
-        Thread.sleep(2000);
-        assertTrue("remove succeeds", constructor.remove(randStr, "Search Suggestions"));
-    }
-
    @Test
     public void removeBatchShouldReturn() throws Exception {
         ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
         assertTrue("batch removal succeeds", constructor.removeBatch(new String[]{this.getRandString(), this.getRandString(), this.getRandString()}, "Search Suggestions"));
-    }
-
-    @Test
-    public void removeConstructorItemShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        constructor.add(new ConstructorItem(randStr), "Search Suggestions");
-        Thread.sleep(2000);
-        assertTrue("remove w/ ConstructorItem succeeds", constructor.remove(new ConstructorItem(randStr), "Search Suggestions"));
     }
 
     @Test
@@ -199,53 +92,4 @@ public class ConstructorTest {
         params.put("suggested_score", 1337);
         assertTrue("modify succeeds", constructor.modify(randStr, randStr, "Search Suggestions", params));
     }
-
-    @Test
-    public void conversionNoParamsShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        assertTrue("conversion without params succeeds", constructor.trackConversion("Stanley_Steamer", "Search Suggestions"));
-    }
-
-    @Test
-    public void conversionWithParamsShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        constructor.add(randStr, "Search Suggestions");
-        Thread.sleep(2000);
-        HashMap<String, Object> params = new HashMap<String, Object>();
-        params.put("item", randStr);
-        assertTrue("conversion with params succeeds", constructor.trackConversion("Stanley_Steamer", "Search Suggestions", params));
-    }
-
-    @Test
-    public void searchNoParamsShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        assertTrue("search without params succeeds", constructor.trackSearch("Stanley_Steamer"));
-    }
-
-    @Test
-    public void searchWithParamsShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        HashMap<String, Object> params = new HashMap<String, Object>();
-        params.put("num_results", 1337);
-        assertTrue("search with params succeeds", constructor.trackSearch("Stanley_Steamer", params));
-    }
-
-    @Test
-    public void clickThroughNoParamsShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        assertTrue("search without params succeeds", constructor.trackClickThrough("Stanley_Steamer", "Search Suggestions"));
-    }
-
-    @Test
-    public void clickThroughWithParamsShouldReturn() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-        String randStr = this.getRandString();
-        constructor.add(randStr, "Search Suggestions");
-        Thread.sleep(2000);
-        HashMap<String, Object> params = new HashMap<String, Object>();
-        params.put("item", randStr);
-        assertTrue("click-through with params succeeds", constructor.trackClickThrough("Stanley_Steamer", "Search Suggestions", params));
-    }
-
 }
